内部ID・外部ID対応とキーワード検索機能を追加

+++ src/在庫管理システム/InventoryCLI.java
@@ -58,7 +58,7 @@ public class InventoryCLI {
 		repo.loadFromFile();
 
 		while (true) {
-			System.out.println("1:商品一覧　2:商品追加　3:在庫追加　4:在庫削減　5:商品検索　6:商品名変更　7:商品削除　8:在庫切れの商品一覧　0:終了");
+			System.out.println("1:商品一覧　2:商品追加　3:在庫の追加・削減　4:キーワード検索　5:商品名変更　6:商品削除　7:在庫切れの商品一覧　0:終了");
 
 			String choiceStr = sc.nextLine();
 			int choice;
@@ -81,32 +81,27 @@ public class InventoryCLI {
 				addProduct();
 				break;
 
-			case 3: //在庫追加
+			case 3: //在庫の追加・削減
 
-				addStock();
+				changeStock();
 				break;
 
-			case 4: //在庫削減
+			case 4: //キーワード検索
 
-				removeStock();
+				keywordSearch();
 				break;
 
-			case 5: //商品検索
-
-				productSearch();
-				break;
-
-			case 6: //商品名変更
+			case 5: //商品名変更
 
 				changeName();
 				break;
 
-			case 7: //商品削除
+			case 6: //商品削除
 
 				removeProduct();
 				break;
 
-			case 8:
+			case 7:
 
 				System.out.println("===在庫切れの商品一覧===");
 				repo.showNoStockProduct();
@@ -129,6 +124,8 @@ public class InventoryCLI {
 		System.out.println("===商品追加===");
 
 		while (true) {
+
+			String id = inputString("ID");
 
 			String name = inputString("商品名");
 			if (repo.productExist(name)) {
@@ -159,7 +156,7 @@ public class InventoryCLI {
 
 			if (confirm("この内容で追加しますか？")) {
 				System.out.println("商品を追加しました");
-				repo.addProduct(name, price, addStock);
+				repo.addProduct(id, name, price, addStock);
 				break;
 			}
 
@@ -175,8 +172,8 @@ public class InventoryCLI {
 
 	}
 
-	private void addStock() {
-		System.out.println("===在庫追加===");
+	private void changeStock() {
+		System.out.println("===在庫の追加・削減===");
 
 		while (true) {
 
@@ -186,8 +183,35 @@ public class InventoryCLI {
 				continue;
 			}
 
-			int amount = inputInt("追加数を入力してください");
-			if (!confirm("商品名:" + name + "   追加数:" + amount + "でよろしいですか？")) {
+			if (confirm("在庫の増加を行う場合はyを、在庫の削減を行う場合はnを入力してください")) {
+
+				//在庫の増加を選んだ場合
+				int amountAdd = inputInt("追加数を入力してください");
+
+				if (!confirm("商品名:" + name + "   追加数:" + amountAdd + "でよろしいですか？")) {
+					if (confirm("入力をやり直しますか？")) {
+						System.out.println("入力をやり直します");
+						continue;
+					} else {
+						System.out.println("選択画面に戻ります");
+						break;
+					}
+				}
+
+				if (amountAdd > 0) {
+					if (repo.findProduct(name).addStock(amountAdd)) {
+						System.out.println("在庫を追加しました");
+						break;
+					} else {
+						System.out.println("0より大きい値を入れてください");
+						continue;
+					}
+				}
+			}
+
+			//在庫の削減を選んだ場合
+			int amountRemove = inputInt("削減数を入力してください");
+			if (!confirm("商品名:" + name + "   削減数:" + amountRemove + "でよろしいですか？")) {
 				if (confirm("入力をやり直しますか？")) {
 					System.out.println("入力をやり直します");
 					continue;
@@ -197,67 +221,66 @@ public class InventoryCLI {
 				}
 			}
 
-			if (repo.findProduct(name).addStock(amount)) {
-				System.out.println("在庫を追加しました");
-				break;
-			} else {
-				System.out.println("0より大きい値を入れてください");
-				continue;
+			if (amountRemove > 0) {
+				if (repo.findProduct(name).removeStock(amountRemove)) {
+					System.out.println("在庫を削減しました");
+					break;
+				} else {
+					System.out.println("在庫不足または0以下の値は指定できません");
+					if (confirm("入力をやり直しますか？")) {
+						System.out.println("入力をやり直します");
+						continue;
+					} else {
+						System.out.println("選択画面に戻ります");
+						break;
+					}
+				}
 			}
+
 		}
 	}
 
-	private void removeStock() {
-		System.out.println("===在庫削減===");
-
+	private void keywordSearch() {
+		System.out.println("===キーワード検索===");
 		while (true) {
+			if (confirm("商品名から検索する場合はyを、IDから検索する場合はnを入力してください")) {
 
-			String name = inputString("商品名");
-			if (!repo.productExist(name)) {
-				System.out.println("入力された名前の商品は存在しません");
+				//商品名からの検索
+				System.out.println("商品名から検索を行います");
+				String keywordName = inputString("キーワード");
+				if (repo.productNameKeywordExist(keywordName)) {
+					System.out.println("キーワードに一致する商品が見つかりました");
+					repo.findProductByNameKeyword(keywordName);
+					break;
+				}
+
+				System.out.println("キーワードに一致する商品は見つかりませんでした");
+				if (confirm("入力をやり直しますか？")) {
+					System.out.println("入力をやり直します");
+					continue;
+				} else {
+					System.out.println("選択画面に戻ります");
+					break;
+				}
+			}
+
+			//IDからの検索
+			System.out.println("IDから検索を行います");
+			String keywordID = inputString("キーワード");
+			if (repo.productIdKeywordExist(keywordID)) {
+				System.out.println("キーワードに一致する商品が見つかりました");
+				repo.findProductByIdKeyword(keywordID);
+			}
+
+			System.out.println("キーワードに一致する商品は見つかりませんでした");
+			if (confirm("入力をやり直しますか？")) {
+				System.out.println("入力をやり直します");
 				continue;
-			}
-
-			int amount = inputInt("削減数を入力してください");
-			if (!confirm("商品名:" + name + "   削減数:" + amount + "でよろしいですか？")) {
-				if (confirm("入力をやり直しますか？")) {
-					System.out.println("入力をやり直します");
-					continue;
-				} else {
-					System.out.println("選択画面に戻ります");
-					break;
-				}
-			}
-			if (repo.findProduct(name).removeStock(amount)) {
-				System.out.println("在庫を削減しました");
-				break;
 			} else {
-				System.out.println("在庫不足または0以下の値は指定できません");
+				System.out.println("選択画面に戻ります");
+				break;
 			}
 		}
-	}
-
-	private void productSearch() {
-		System.out.println("===商品検索===");
-		while (true) {
-			String name = inputString("商品名");
-			if (!repo.productExist(name)) {
-				System.out.println("入力された名前の商品は存在しません");
-				if (confirm("入力をやり直しますか？")) {
-					System.out.println("入力をやり直します");
-					continue;
-				} else {
-					System.out.println("選択画面に戻ります");
-					break;
-				}
-			}
-
-			Product p = repo.findProduct(name);
-
-			System.out.println(p);
-			break;
-		}
-
 	}
 
 	private void changeName() {
@@ -295,8 +318,13 @@ public class InventoryCLI {
 				}
 			}
 
+			if (!confirm("本当に削除してもよろしいですか？")) {
+				System.out.println("選択画面に戻ります");
+				break;
+			}
 			repo.removeProduct(name);
 			System.out.println("商品を削除しました");
+
 			break;
 		}
 	}
